pyproject.toml
[tool.poetry]
name = "watchr"
version = "0.2.3"
description = "Picking up and processing files"
authors = ["Bryan Jenks <bryan@bryanjenks.dev>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.0.1"
loguru = "^0.7.0"
watchdog = "^3.0.0"
psycopg2-binary = "^2.9.5"
fastapi = "^0.95.1"
FastAPI-SQLAlchemy = "^0.2.1"
pydantic = "^1.10.7"
alembic = "^1.10.4"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
pretty-errors = "^1.2.25"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.3.1"

[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | docs
  | notes
)/
'''

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
skip_glob = [ "docs", "notes", "img" ]
known_third_party = ["loguru", "pytest"]

[tool.vulture]
exclude = []
ignore_decorators = ["@app.route", "@require_*"]
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["src/"]
sort_by_size = true
verbose = false

[tool.interrogate]
ignore-init-method = false
ignore-init-module = false
ignore-magic = false
ignore-semiprivate = false
ignore-private = false
ignore-property-decorators = false
ignore-module = false
ignore-nested-functions = false
ignore-nested-classes = false
ignore-setters = false
fail-under = 0
exclude = ["setup.py", "docs", "build", "src"]
ignore-regex = ["^get$", "^mock_.*", ".*BaseClass.*"]
verbose = 2
quiet = false
whitelist-regex = []
color = true
generate-badge = "images/interrogate_badge.svg"
badge-format = "svg"
